figures/requirements.txt
biopython
glasbey
hdbscan
matplotlib
numpy
pandas
python-dotenv
scikit-learn
scipy
tqdm
umap-learn